+++ src/test/java/racingcar/views/InputViewTest.java
@@ -1,7 +1,6 @@
 package racingcar.views;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import camp.nextstep.edu.missionutils.Console;
@@ -13,9 +12,7 @@ import org.junit.jupiter.api.Test;
 import racingcar.domain.Car;
 import racingcar.domain.RacingCars;
 import racingcar.exception.CarNameDuplicateException;
-import racingcar.exception.CarNameIncorrectException;
 import racingcar.exception.CarNameSizeLimitExceededException;
-import racingcar.utils.InputValidate;
 
 class InputViewTest {
 
@@ -33,35 +30,6 @@ class InputViewTest {
         String result = Console.readLine();
 
         assertThat(result).isEqualTo(input);
-    }
-
-    @Test
-    @DisplayName("자동차 이름 입력 시 빈값, null 입력 시 예외 발생")
-    void nullAndEmptyCheckTest() {
-
-        String inputNull = null;
-        String inputEmpty = "";
-        String inputBlank = "  ";
-
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateCarNames(inputNull);
-        });
-        assertThrows(CarNameIncorrectException.class, () -> {
-            InputValidate.validateCarNames(inputNull);
-        });
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateCarNames(inputEmpty);
-        });
-        assertThrows(CarNameIncorrectException.class, () -> {
-            InputValidate.validateCarNames(inputEmpty);
-        });
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateCarNames(inputBlank);
-        });
-        assertThrows(CarNameIncorrectException.class, () -> {
-            InputValidate.validateCarNames(inputBlank);
-        });
-
     }
 
     @Test
@@ -91,7 +59,7 @@ class InputViewTest {
 
         RacingCars racingCars = InputView.getRacingCars();
         List<Car> carNameList = racingCars.getRacingCars();
-        
+
         assertThat(carNameList.size()).isEqualTo(3);
         assertThat(carNameList).isNotNull();
     }
@@ -105,34 +73,6 @@ class InputViewTest {
         // console.readLine() 에 테스트 입력 값 등록
         System.setIn(new ByteArrayInputStream(error.getBytes()));
         assertThrows(CarNameSizeLimitExceededException.class, InputView::getRacingCars);
-    }
-
-    @Test
-    @DisplayName("도전 횟수 입력 받기 숫자 타입 체크")
-    void tryCountInputAndNumberTypeCheckTest() {
-        // 정상의 경우
-        assertDoesNotThrow(() -> InputValidate.validateMovingCount("123"));
-        // 비정상의 경우 예외 발생
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateMovingCount(" ");
-        });
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateMovingCount("1 3");
-        });
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateMovingCount(" a");
-        });
-        // 0 이하의 값 입려 시 예외 발생
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateMovingCount("0");
-        });
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateMovingCount("-1");
-        });
-        // Integer Max 값 이상의 경우 예외 발생
-        assertThrows(IllegalArgumentException.class, () -> {
-            InputValidate.validateMovingCount("2147483648");
-        });
     }
 
     @Test
